+++ src/main/java/org/oracleone/forohub/service/UserService.java
@@ -3,6 +3,7 @@ import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.oracleone.forohub.enums.Role;
 import org.oracleone.forohub.persistence.DTO.UserDTO;
+import org.oracleone.forohub.persistence.DTO.UserDetailsDTO;
 import org.oracleone.forohub.persistence.DTO.UserRegisterDTO;
 import org.oracleone.forohub.persistence.repositories.UserRepository;
 import org.oracleone.forohub.utils.UserConverter;
@@ -10,13 +11,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.oracleone.forohub.persistence.entities.User;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.stream.Collectors;
 
 @Service
-public class UserService{
+public class UserService {
 
     private final UserRepository userRepository;
     private final UserConverter userConverter;
@@ -30,15 +34,15 @@ public class UserService{
     }
 
     @Transactional
-    public void saveUser(UserRegisterDTO userRegisterDTO){
+    public void saveUser(UserRegisterDTO userRegisterDTO) {
         this.userRepository.save(userRegisterDTOToUser(userRegisterDTO));
     }
 
-    public User getUserById(Long id){
+    public User getUserById(Long id) {
         return this.userRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("User not found"));
     }
 
-    public User userRegisterDTOToUser(UserRegisterDTO userRegisterDTO){
+    public User userRegisterDTOToUser(UserRegisterDTO userRegisterDTO) {
         return new User(
                 userRegisterDTO.name(),
                 userRegisterDTO.email(),
@@ -48,7 +52,7 @@ public class UserService{
         );
     }
 
-    public Page<UserDTO> getAllUsers(Pageable pageable){
+    public Page<UserDTO> getAllUsers(Pageable pageable) {
         Page<User> users = userRepository.findAll(pageable);
         if (users.isEmpty()) {
             throw new EntityNotFoundException("No users found");
@@ -57,18 +61,49 @@ public class UserService{
                 this.userConverter::EntityToDTO
         );
     }
-
+    /* MARKED TO DELETE
     public User getByNameAndEmail(String name, String email){
         return this.userRepository.findByNameAndEmail(name,email).orElseThrow(
                 EntityNotFoundException::new
         );
-    }
+    }*/
 
-    public UserDTO convertToDTO(User user){
+    public UserDTO convertToDTO(User user) {
         return this.userConverter.EntityToDTO(user);
     }
 
+    @Transactional
     public void deleteUser(Long id) {
-        this.userRepository.deleteById(id);
+        User user = this.userRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("User not found"));
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        User authenticatedUser = findByEmail(userDetails.getUsername());
+
+        if (!user.equals(authenticatedUser)) {
+            throw new AccessDeniedException("You are not authorized to delete this User");
+        } else {
+            this.userRepository.deleteById(id);
+        }
+    }
+
+    //TASK : ADD UPDATE METHOD
+    @Transactional
+    public void updateUser(UserDetailsDTO userDetailsDTO, Long id) {
+        User user = this.userRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("User not found"));
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        User authenticatedUser = findByEmail(userDetails.getUsername());
+
+        if (!user.equals(authenticatedUser)) {
+            throw new AccessDeniedException("You are not authorized to update this User");
+        } else {
+            user.setName(userDetailsDTO.name());
+            user.setEmail(userDetailsDTO.email());
+            user.setPassword(passwordEncoder.encode(userDetailsDTO.password()));
+            this.userConverter.EntityToDTO(this.userRepository.save(user));
+        }
+    }
+
+    public User findByEmail(String email) {
+        return this.userRepository.findByEmail(email).orElseThrow(() -> new EntityNotFoundException("User not found"));
     }
 }
+
